requirements.txt
flickrapi
requests
python-decouple
alive-progress
flake8
